Added experimental time mapping functionality
Layers and Shapefiles now can be time-stamped so that each shape is
defined by certain time-fields and can be read with a shape's GetTime()
method.

NewMap now has a SaveTimeSequence method that uses the time-stamp of
each shape to automatically create a folder of map images that change
across time.